+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Replace `cl` with `cl-lib`
 - Edit process buffer in place to avoid multibyte flag resetting
 - Allow binding `url-mime-accept-string`
+- Respect `url-request-noninteractive`
 
 ## [0.9.0] - 2021-12-05 ##